Task: refactor src and tests (#76)
* refactored src and test

* removed testing file
